Notification Stuff Part 3v3 (finished! Can send notif to both signed in and checked in attendees. NO DUPLICATION IN NOTIFICATION) - EventDetailsActivity -- removed notification stuff like sendNotification() and moved them into PushNotificationService.java - PushNotificationService.java -- added notification stuff like sendNotification() from EventDetailsActivity.java so i can just call this class to use sendNotification() - Database.java -- added Notification collection -- added a function that stores the Notification message - QRScanningActivity.java -- When the attendee checks in, automatically subscribe the attendee to recieve notifications from the organizer via the id of the event. If the user is in both signed in and checked in, firebase handles the duplicate notifications.
