+++ code/app/src/main/java/com/example/emojibrite/Database.java
@@ -28,12 +28,14 @@ import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.QueryDocumentSnapshot;
 import com.google.firebase.firestore.SetOptions;
+import com.google.firebase.firestore.Transaction;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.ListResult;
 import com.google.firebase.storage.StorageMetadata;
 import com.google.firebase.storage.StorageReference;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -53,7 +55,7 @@ public class Database {
 
     private final CollectionReference signedUpRef = db.collection("SignedUp");
 
-
+    private final CollectionReference notificationRef = db.collection("Notifications");
 
     private String firestoreDebugTag = "Firestore";
 
@@ -766,5 +768,37 @@ once created, u can call getuseruid to get the user id and use it to get user da
             Log.e(TAG, "Error updating event geolocation check-ins", e);
         });
     }
+
+    // Notification Database Section //
+    public void storeNotification(String message, String eventId) {
+        // Get a reference to the 'Notifications' collection
+
+        // Get a reference to the document that represents the event
+        DocumentReference eventDocRef = notificationRef.document(eventId);
+
+        // Use the Firestore transaction feature to safely update the array
+        db.runTransaction((Transaction.Function<Void>) transaction -> {
+                    // Get the current state of the document
+                    DocumentSnapshot snapshot = transaction.get(eventDocRef);
+
+                    // If the document does not exist, create a new document with the message
+                    if (!snapshot.exists()) {
+                        Map<String, Object> data = new HashMap<>();
+                        data.put("messages", Arrays.asList(message));
+                        transaction.set(eventDocRef, data);
+                    } else {
+                        // If the document exists, add the new message to the array of messages
+                        List<String> messages = (List<String>) snapshot.get("messages");
+                        messages.add(message);
+                        transaction.update(eventDocRef, "messages", messages);
+                    }
+
+                    // Return null because the function is of type Void
+                    return null;
+                }).addOnSuccessListener(aVoid -> Log.d(TAG, "Notification successfully written!"))
+                .addOnFailureListener(e -> Log.w(TAG, "Error writing notification", e));
+    }
+
+    // Notification Database Section //
 
 }
